requirements.txt
fastapi
uvicorn[standard]
httpx
psycopg[binary]
SQLAlchemy
alembic
pydantic
pydantic-settings
python-dotenv
paramiko
psycopg2-binary
